:ambulance: Remove obsolete --force flag from mike
In the latest v2.0.0 of mike the --force flag is removed, which is a
breaking change. Which consequently has broken the github workflow for
building the schism documentations too. The changelog of mike related to
this breaking-change is posted here https://github.com/jimporter/mike/commit/44cc579296327da713a18f93ea5e4fa5a6e8a715